requirements.txt
Django==4.1.5
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django_debug_toolbar==3.8.1
django-cors-headers==3.13.0
requests==2.28.2
pytest-django==4.5.2
pytest-xdist==3.2.0
pytest-mock==3.10.0
model-bakery==1.10.1
flake8==6.0.0
shortuuid==1.0.11
gunicorn==20.1.0
redis==4.4.2
celery==5.2.7
django-redis==5.2.0
aiohttp==3.8.4